"久违的提交 #2"

+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/bangumi/info/index/remote/FirstEp.kt
@@ -2,5 +2,5 @@ package cn.spacexc.bilibilisdk.sdk.bangumi.info.index.remote
 
 data class FirstEp(
     val cover: String,
-    val ep_id: Int
+    val ep_id: Long
 )
